add Decimal

+++ lib/common/src/test/java/study/daydayup/wolf/common/model/currency/RMBTest.java
@@ -18,22 +18,22 @@ public class RMBTest {
     public void toYuan() {
         RMB r = new RMB(3, RMBEnum.YUAN);
 
-        long vY = r.toYuan();
+        long vY = r.toYuan().toLong();
         assertEquals("convert to yuan fail", 3, vY);
 
-        long vJ = r.toJiao();
+        long vJ = r.toJiao().toLong();
         assertEquals("convert to jiao fail", 30, vJ);
 
-        long vF = r.toFen();
+        long vF = r.toFen().toLong();
         assertEquals("convert to fen fail", 300, vF);
 
-        long vL = r.toLi();
+        long vL = r.toLi().toLong();
         assertEquals("convert to li fail", 3000, vL);
 
-        long vH = r.toHao();
+        long vH = r.toHao().toLong();
         assertEquals("convert to hao fail", 30000, vH);
 
-        long vS = r.toSi();
+        long vS = r.toSi().toLong();
         assertEquals("convert to si fail", 300000, vS);
 
     }
@@ -42,22 +42,22 @@ public class RMBTest {
     public void toSi() {
         RMB r = new RMB(300000, RMBEnum.SI);
 
-        long vY = r.toYuan();
+        long vY = r.toYuan().toLong();
         assertEquals("convert to yuan fail", 3, vY);
 
-        long vJ = r.toJiao();
+        long vJ = r.toJiao().toLong();
         assertEquals("convert to jiao fail", 30, vJ);
 
-        long vF = r.toFen();
+        long vF = r.toFen().toLong();
         assertEquals("convert to fen fail", 300, vF);
 
-        long vL = r.toLi();
+        long vL = r.toLi().toLong();
         assertEquals("convert to li fail", 3000, vL);
 
-        long vH = r.toHao();
+        long vH = r.toHao().toLong();
         assertEquals("convert to hao fail", 30000, vH);
 
-        long vS = r.toSi();
+        long vS = r.toSi().toLong();
         assertEquals("convert to si fail", 300000, vS);
 
     }
@@ -66,22 +66,22 @@ public class RMBTest {
     public void toFen() {
         RMB r = new RMB(300, RMBEnum.FEN);
 
-        long vY = r.toYuan();
+        long vY = r.toYuan().toLong();
         assertEquals("convert to yuan fail", 3, vY);
 
-        long vJ = r.toJiao();
+        long vJ = r.toJiao().toLong();
         assertEquals("convert to jiao fail", 30, vJ);
 
-        long vF = r.toFen();
+        long vF = r.toFen().toLong();
         assertEquals("convert to fen fail", 300, vF);
 
-        long vL = r.toLi();
+        long vL = r.toLi().toLong();
         assertEquals("convert to li fail", 3000, vL);
 
-        long vH = r.toHao();
+        long vH = r.toHao().toLong();
         assertEquals("convert to hao fail", 30000, vH);
 
-        long vS = r.toSi();
+        long vS = r.toSi().toLong();
         assertEquals("convert to si fail", 300000, vS);
     }
 }
